feat: add bar chart for data visualization

+++ NutritionApp/Components/Charting/DataSet.cs
@@ -4,6 +4,6 @@ public class Dataset
     public string? Label { get; set; }
     public string? BackgroundColor { get; set; }
     public string? BorderColor { get; set; }
-    public int[]? Data { get; set; }
+    public double[]? Data { get; set; }
     public int BorderWidth { get; set; }
 }
